Make video importing an effect
