+++ exe/Main.hs
@@ -3,7 +3,7 @@
 import           Komposition.Prelude
 
 import           Komposition.Application
-import           Komposition.Import.Video
+import           Komposition.Import.Video.FFmpeg
 import           Komposition.Logging.FastLogger
 import           Komposition.Project.Store.File
 import           Komposition.UserInterface.GtkInterface
@@ -15,6 +15,5 @@ main = do
   initialize
   cssPath   <- getDataFileName "style.css"
   loggerSet <- newStderrLoggerSet 1024
-  runGtkUserInterface cssPath
-                      (runFastLoggerLog loggerSet . runFileProjectStoreIO)
-                      komposition
+  let runEffects = runFastLoggerLog loggerSet . runFileProjectStoreIO . runFFmpegVideoImport
+  runGtkUserInterface cssPath runEffects komposition
