+++ src/Komposition/Application.hs
@@ -21,11 +21,13 @@ import           Data.Row.Records
 
 import           Komposition.Application.KeyMaps
 import           Komposition.Application.WelcomeScreenMode
+import           Komposition.Import.Video
 import           Komposition.Project.Store
 
 komposition
   :: ( Application t m sig
     , Member ProjectStore sig
+    , Member VideoImport sig
     , Carrier sig m
     )
   => t m Empty Empty ()
